+++ lib/pack_idea/pack_idea.dart
@@ -18,8 +18,10 @@ import 'package:rxdart/rxdart.dart';
 
 part 'screens/create_idea_screen.dart';
 part 'screens/idea_answer_screen.dart';
+part 'screens/idea_answer_screen_state.dart';
 part 'screens/idea_screen.dart';
-part 'state/idea_questions_initializer.dart';
+part 'screens/idea_screen_state.dart';
+part 'states/idea_questions_initializer.dart';
 part 'widgets/answer_creator.dart';
 part 'widgets/answer_field.dart';
 part 'widgets/answer_field_bubble.dart';