+++ lib/pack_idea/screens/idea_answer_screen.dart
@@ -26,14 +26,6 @@ class IdeaAnswerScreen extends HookWidget {
     return null;
   }
 
-  void back({
-    required final BuildContext context,
-    required final IdeaProject idea,
-  }) {
-    closeKeyboard(context: context);
-    onBack(idea);
-  }
-
   @override
   Widget build(final BuildContext context) {
     final ideasProvider = context.read<IdeaProjectsProvider>();
@@ -45,26 +37,13 @@ class IdeaAnswerScreen extends HookWidget {
     // ignore: close_sinks
     final updatesStream = useStreamController<bool>();
 
-    textController.addListener(() {
-      if (answer.value.text == textController.text) return;
-      answer.value.text = textController.text;
-      maybeIdea.updated = DateTime.now();
-      updatesStream.add(true);
-    });
-
-    updatesStream.stream
-        .throttleTime(
-      const Duration(milliseconds: 700),
-      leading: true,
-      trailing: true,
-    )
-        .forEach(
-      (final _) async {
-        maybeIdea.folder?.sortProjectsByDate(project: maybeIdea);
-        // ideaProjectsProvider.notify();
-        await answer.value.save();
-        await maybeIdea.save();
-      },
+    final state = useIdeaAnswerScreenState(
+      answer: answer,
+      context: context,
+      idea: maybeIdea,
+      onScreenBack: onBack,
+      textController: textController,
+      updatesStream: updatesStream,
     );
 
     return Scaffold(
@@ -83,7 +62,7 @@ class IdeaAnswerScreen extends HookWidget {
             ),
           ),
         ),
-        onBack: () => back(idea: maybeIdea, context: context),
+        onBack: state.onBack,
       ),
       body: Center(
         child: Padding(
@@ -105,7 +84,7 @@ class IdeaAnswerScreen extends HookWidget {
                       filled: false,
                       endlessLines: true,
                       focusOnInit: textController.text.isEmpty,
-                      onSubmit: () => back(idea: maybeIdea, context: context),
+                      onSubmit: state.onBack,
                       controller: textController,
                     ),
                   ),