+++ lib/pack_note/screens/note_project_screen.dart
@@ -1,28 +1,25 @@
-part of note_project;
+part of pack_note;
 
 class NoteProjectScreen extends HookWidget {
   const NoteProjectScreen({
     required final this.noteId,
     required final this.onBack,
+    required this.onGoHome,
+    required this.checkIsProjectActive,
     final Key? key,
   }) : super(key: key);
   final String noteId;
   final ValueChanged<NoteProject> onBack;
-
-  void back({
-    required final BuildContext context,
-    required final NoteProject note,
-  }) {
-    closeKeyboard(context: context);
-    onBack(note);
-  }
+  final BoolValueChanged<BasicProject> checkIsProjectActive;
+  final VoidCallback onGoHome;
 
   @override
   Widget build(final BuildContext context) {
     final theme = Theme.of(context);
-    final silentFolderProvider = context.read<FolderStateProvider>();
     final screenLayout = ScreenLayout.of(context);
+
     final noteFocusNode = useFocusNode();
+
     final noteProvider = context.read<NoteProjectsProvider>();
     final maybeNote = noteProvider.state[noteId]!;
 
@@ -30,37 +27,17 @@ class NoteProjectScreen extends HookWidget {
     final noteController = useTextEditingController(text: maybeNote.note);
 
     // ignore: close_sinks
-    final updatesStream = useStreamController<bool>();
-    noteController.addListener(() {
-      if (note.value.note == noteController.text) return;
-      bool updateFolder = false;
-      if (note.value.title != NoteProject.getTitle(noteController.text)) {
-        updateFolder = true;
-      } else {
-        updateFolder = note.value.folder?.projectsList.first != note.value;
-      }
-      note.value
-        ..note = noteController.text
-        ..updated = DateTime.now();
-      updatesStream.add(updateFolder);
-    });
+    final updatesStream = useStreamController<NoteProjectNotifier>();
 
-    updatesStream.stream
-        .throttleTime(
-      const Duration(milliseconds: 700),
-      leading: true,
-      trailing: true,
-    )
-        .forEach((final updateFolder) async {
-      noteProvider.put(key: note.value.id, value: note.value);
-
-      if (updateFolder) {
-        note.value.folder?.sortProjectsByDate(project: note.value);
-        silentFolderProvider.notify();
-      }
-
-      return note.value.save();
-    });
+    final state = useNoteProjectScreenState(
+      context: context,
+      note: note.value,
+      onScreenBack: onBack,
+      noteController: noteController,
+      updatesStream: updatesStream,
+      onGoHome: onGoHome,
+      checkIsProjectActive: checkIsProjectActive,
+    );
 
     return Scaffold(
       backgroundColor: theme.canvasColor,
@@ -70,7 +47,15 @@ class NoteProjectScreen extends HookWidget {
         height: screenLayout.small ? null : 30,
         screenLayout: screenLayout,
         titleStr: '',
-        onBack: () => back(context: context, note: note.value),
+        // actions: [
+        //   if (!isDesktop)
+        //     CupertinoIconButton(
+        //       onPressed: state.onSettings,
+        //       icon: Icons.more_vert_rounded,
+        //     ),
+        //   const SizedBox(width: 20),
+        // ],
+        onBack: state.onBack,
       ),
       body: Center(
         child: ConstrainedBox(
@@ -92,16 +77,21 @@ class NoteProjectScreen extends HookWidget {
                         hintText: S.current.writeANote,
                         fillColor: Colors.transparent,
                         filled: false,
+                        limit: note.value.charactersLimit,
                         focusNode: noteFocusNode,
                         endlessLines: true,
-                        onSubmit: () =>
-                            back(context: context, note: note.value),
+                        onSubmit: state.onBack,
                         controller: noteController,
                       ),
                     ),
                     Column(
                       mainAxisAlignment: MainAxisAlignment.end,
                       children: [
+                        NoteSettingsButton(
+                          note: note.value,
+                          onRemove: state.onRemove,
+                          updatesStream: updatesStream,
+                        ),
                         SpecialEmojiPopup(
                           controller: noteController,
                           focusNode: noteFocusNode,